// src/lib/models/AuthResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AuthResponse = {
    userId?: string;
    username?: string;
    email?: string;
    role?: string;
    token?: string;
    premierConnexion?: boolean;
    connectionTime?: number;
    avote?: boolean;
};